This replaces the express and apollo server 1.x with apollo server 2.x
