+++ package.json
@@ -13,15 +13,10 @@
     "lint:fix": "npm run lint -- --fix"
   },
   "dependencies": {
-    "apollo-server-express": "^1.4.0",
-    "body-parser": "^1.19.0",
-    "cuid": "^2.1.8",
-    "express": "^4.17.1",
+    "apollo-server": "^2.9.16",
     "graphql": "^14.5.8",
-    "graphql-tools": "^4.0.6",
     "lodash": "^4.17.15",
-    "mongoose": "^5.8.2",
-    "morgan": "^1.9.1"
+    "mongoose": "^5.8.2"
   },
   "devDependencies": {
     "@shelf/jest-mongodb": "^1.1.3",
